refactor to include model package and lombok
