+++ pom.xml
@@ -57,6 +57,13 @@
 			<artifactId>spring-boot-starter-thymeleaf</artifactId>
 			<version>3.1.3</version> 
 		</dependency>
+        <!-- lombok -->
+        <dependency>
+		    <groupId>org.projectlombok</groupId>
+		    <artifactId>lombok</artifactId>
+		    <version>1.18.30</version>
+		    <scope>provided</scope>
+	    </dependency>
     </dependencies>
 
     <build>
