+++ src/main/java/com/greengroupecommerce/controllers/ProductsController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
 
-import com.greengroupecommerce.Product;
+import com.greengroupecommerce.model.Product;
 
 @Controller
 public class ProductsController {
