+++ src/main/java/com/greengroupecommerce/model/Product.java
@@ -1,0 +1,12 @@
+package com.greengroupecommerce.model;
+
+import lombok.AllArgsConstructor;
+
+@AllArgsConstructor
+public class Product {
+  public String image;
+  public String name;
+  public String description;
+  public String price;
+  
+}
